agent_template_71289/requirements.txt
```
python-dotenv
langchain-nvidia-ai-endpoints
langgraph
langchain-community-tools
langchain-community-utilities
langchain-core
```